refactor: split napi lint sanitize modules
