+++ crates/ox_content_napi/src/sanitize.rs
@@ -1,5 +1,3 @@
-use rustc_hash::FxHashSet;
-
 use crate::JsSanitizeOptions;
 
 pub fn sanitize_html(html: &str, options: Option<&JsSanitizeOptions>) -> String {
@@ -12,184 +10,14 @@ pub fn sanitize_html(html: &str, options: Option<&JsSanitizeOptions>) -> String 
     sanitize_html_with_config(html, &SanitizeConfig::from_js(options))
 }
 
-struct SanitizeConfig {
-    tags: FxHashSet<String>,
-    attributes: FxHashSet<String>,
-    url_schemes: FxHashSet<String>,
-}
+mod config;
+mod parser;
 
-impl Default for SanitizeConfig {
-    fn default() -> Self {
-        Self {
-            tags: [
-                "a",
-                "audio",
-                "blockquote",
-                "br",
-                "code",
-                "del",
-                "details",
-                "div",
-                "em",
-                "h1",
-                "h2",
-                "h3",
-                "h4",
-                "h5",
-                "h6",
-                "hr",
-                "iframe",
-                "img",
-                "input",
-                "li",
-                "nav",
-                "ol",
-                "p",
-                "picture",
-                "pre",
-                "source",
-                "span",
-                "strong",
-                "summary",
-                "sup",
-                "table",
-                "tbody",
-                "td",
-                "th",
-                "thead",
-                "track",
-                "tr",
-                "ul",
-                "video",
-            ]
-            .into_iter()
-            .map(ToString::to_string)
-            .collect(),
-            attributes: [
-                "allow",
-                "allowfullscreen",
-                "alt",
-                "aria-label",
-                "autoplay",
-                "checked",
-                "class",
-                "controls",
-                "controlslist",
-                "crossorigin",
-                "data-code-title",
-                "data-group",
-                "data-line",
-                "data-line-number",
-                "data-line-number-start",
-                "data-line-numbers",
-                "data-ox-tab-group",
-                "default",
-                "disabled",
-                "disablepictureinpicture",
-                "disableremoteplayback",
-                "height",
-                "href",
-                "id",
-                "kind",
-                "label",
-                "loading",
-                "loop",
-                "media",
-                "muted",
-                "name",
-                "playsinline",
-                "poster",
-                "preload",
-                "referrerpolicy",
-                "rel",
-                "sandbox",
-                "sizes",
-                "src",
-                "srcset",
-                "srclang",
-                "style",
-                "target",
-                "title",
-                "type",
-                "width",
-            ]
-            .into_iter()
-            .map(ToString::to_string)
-            .collect(),
-            url_schemes: ["http", "https", "mailto", "tel"]
-                .into_iter()
-                .map(ToString::to_string)
-                .collect(),
-        }
-    }
-}
+#[cfg(test)]
+mod tests;
 
-impl SanitizeConfig {
-    fn from_js(options: &JsSanitizeOptions) -> Self {
-        let mut config = Self::default();
-        if let Some(tags) = &options.allowed_tags {
-            config.tags = tags.iter().map(|value| value.to_ascii_lowercase()).collect();
-        }
-        if let Some(attrs) = &options.allowed_attributes {
-            config.attributes = attrs.iter().map(|value| value.to_ascii_lowercase()).collect();
-        }
-        if let Some(schemes) = &options.allowed_url_schemes {
-            config.url_schemes = schemes.iter().map(|value| value.to_ascii_lowercase()).collect();
-        }
-        config
-    }
-
-    fn allows_tag(&self, tag: &str) -> bool {
-        self.tags.contains(tag)
-    }
-
-    fn allows_attr(&self, attr: &str) -> bool {
-        if attr.starts_with("data-") || attr.starts_with("aria-") {
-            return self.attributes.contains("data-*")
-                || self.attributes.contains("aria-*")
-                || self.attributes.contains(attr);
-        }
-        self.attributes.contains(attr)
-    }
-
-    fn allows_url(&self, value: &str) -> bool {
-        let trimmed = value.trim_matches(|ch: char| ch.is_ascii_control() || ch.is_whitespace());
-        if trimmed.is_empty()
-            || trimmed.starts_with('/')
-            || trimmed.starts_with("./")
-            || trimmed.starts_with("../")
-            || trimmed.starts_with('#')
-        {
-            return true;
-        }
-        let Some(colon) = trimmed.find(':') else {
-            return true;
-        };
-        let first_path_marker = trimmed.find(&['/', '?', '#'][..]).unwrap_or(usize::MAX);
-        if first_path_marker < colon {
-            return true;
-        }
-        let scheme = trimmed[..colon]
-            .chars()
-            .filter(|ch| !ch.is_ascii_whitespace())
-            .flat_map(char::to_lowercase)
-            .collect::<String>();
-        self.url_schemes.contains(&scheme)
-    }
-
-    fn allows_srcset(&self, value: &str) -> bool {
-        value.split(',').all(|candidate| {
-            let candidate = candidate.trim();
-            if candidate.is_empty() {
-                return false;
-            }
-            let url_end =
-                candidate.find(|ch: char| ch.is_ascii_whitespace()).unwrap_or(candidate.len());
-            let url = &candidate[..url_end];
-            !url.is_empty() && self.allows_url(url)
-        })
-    }
-}
+use config::SanitizeConfig;
+use parser::{find_ci, is_attr_name, scan_tag_end, ParsedAttr, ParsedTag};
 
 fn sanitize_html_with_config(html: &str, config: &SanitizeConfig) -> String {
     if !html.contains('<') {
@@ -276,99 +104,6 @@ fn sanitize_html_with_config(html: &str, config: &SanitizeConfig) -> String {
     out
 }
 
-#[derive(Debug)]
-struct ParsedTag<'a> {
-    name: String,
-    closing: bool,
-    self_closing: bool,
-    attrs: Vec<ParsedAttr<'a>>,
-}
-
-#[derive(Debug)]
-struct ParsedAttr<'a> {
-    name: String,
-    value: Option<&'a str>,
-}
-
-impl<'a> ParsedTag<'a> {
-    fn parse(raw: &'a str) -> Option<Self> {
-        let raw = raw.trim();
-        let closing = raw.starts_with('/');
-        let raw = raw.strip_prefix('/').unwrap_or(raw).trim_start();
-        let bytes = raw.as_bytes();
-        let mut cursor = 0usize;
-        while cursor < bytes.len()
-            && !bytes[cursor].is_ascii_whitespace()
-            && bytes[cursor] != b'/'
-            && bytes[cursor] != b'>'
-        {
-            cursor += 1;
-        }
-        if cursor == 0 {
-            return None;
-        }
-        let name = raw[..cursor].to_ascii_lowercase();
-        if !is_tag_name(&name) {
-            return None;
-        }
-        let attr_text = raw[cursor..].trim();
-        let self_closing = attr_text.ends_with('/');
-        let attr_text = attr_text.trim_end_matches('/').trim();
-        Some(Self { name, closing, self_closing, attrs: parse_attrs(attr_text) })
-    }
-}
-
-fn parse_attrs(mut raw: &str) -> Vec<ParsedAttr<'_>> {
-    let mut attrs = Vec::new();
-    while !raw.is_empty() {
-        raw = raw.trim_start();
-        if raw.is_empty() {
-            break;
-        }
-        let bytes = raw.as_bytes();
-        let mut name_end = 0usize;
-        while name_end < bytes.len()
-            && !bytes[name_end].is_ascii_whitespace()
-            && bytes[name_end] != b'='
-            && bytes[name_end] != b'/'
-        {
-            name_end += 1;
-        }
-        if name_end == 0 {
-            break;
-        }
-        let name = raw[..name_end].to_ascii_lowercase();
-        raw = &raw[name_end..];
-        raw = raw.trim_start();
-        let mut value = None;
-        if raw.starts_with('=') {
-            raw = raw[1..].trim_start();
-            if let Some(quote) =
-                raw.as_bytes().first().copied().filter(|b| *b == b'"' || *b == b'\'')
-            {
-                let value_start = 1usize;
-                if let Some(end) = raw[value_start..].bytes().position(|byte| byte == quote) {
-                    let value_end = value_start + end;
-                    value = Some(&raw[value_start..value_end]);
-                    raw = &raw[value_end + 1..];
-                } else {
-                    value = Some(&raw[value_start..]);
-                    raw = "";
-                }
-            } else {
-                let value_end = raw
-                    .bytes()
-                    .position(|byte| byte.is_ascii_whitespace() || byte == b'/')
-                    .unwrap_or(raw.len());
-                value = Some(&raw[..value_end]);
-                raw = &raw[value_end..];
-            }
-        }
-        attrs.push(ParsedAttr { name, value });
-    }
-    attrs
-}
-
 fn write_sanitized_attr(out: &mut String, attr: &ParsedAttr<'_>, config: &SanitizeConfig) {
     if attr.name.starts_with("on") || !is_attr_name(&attr.name) || !config.allows_attr(&attr.name) {
         return;
@@ -387,51 +122,6 @@ fn write_sanitized_attr(out: &mut String, attr: &ParsedAttr<'_>, config: &Saniti
         escape_html_attr(value, out);
         out.push('"');
     }
-}
-
-fn scan_tag_end(html: &str, start: usize) -> Option<usize> {
-    let bytes = html.as_bytes();
-    let mut cursor = start;
-    let mut quote = None;
-    while cursor < bytes.len() {
-        match quote {
-            Some(q) if bytes[cursor] == q => quote = None,
-            Some(_) => {}
-            None if bytes[cursor] == b'"' || bytes[cursor] == b'\'' => quote = Some(bytes[cursor]),
-            None if bytes[cursor] == b'>' => return Some(cursor + 1),
-            None => {}
-        }
-        cursor += 1;
-    }
-    None
-}
-
-fn find_ci(haystack: &str, from: usize, needle: &str) -> Option<usize> {
-    let hay = haystack.as_bytes();
-    let pat = needle.as_bytes();
-    if pat.is_empty() || from > hay.len() || hay.len() - from < pat.len() {
-        return None;
-    }
-    let last_start = hay.len() - pat.len();
-    let rest = &pat[1..];
-    let mut base = from;
-    while base <= last_start {
-        let rel = memchr::memchr(pat[0], &hay[base..=last_start])?;
-        let index = base + rel;
-        if hay[index + 1..index + pat.len()].eq_ignore_ascii_case(rest) {
-            return Some(index);
-        }
-        base = index + 1;
-    }
-    None
-}
-
-fn is_tag_name(value: &str) -> bool {
-    value.bytes().all(|byte| byte.is_ascii_alphanumeric() || byte == b'-')
-}
-
-fn is_attr_name(value: &str) -> bool {
-    value.bytes().all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b':' | b'_'))
 }
 
 fn escape_html_text(value: &str, out: &mut String) {
@@ -454,63 +144,5 @@ fn escape_html_attr(value: &str, out: &mut String) {
             '>' => out.push_str("&gt;"),
             _ => out.push(ch),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn removes_scripts_and_event_handlers() {
-        let html =
-            r#"<p onclick="x()">Hi<script>alert(1)</script><a href="javascript:x">x</a></p>"#;
-        let sanitized = sanitize_html(html, Some(&JsSanitizeOptions::default()));
-
-        assert!(!sanitized.contains("script"));
-        assert!(!sanitized.contains("onclick"));
-        assert!(!sanitized.contains("javascript"));
-        assert!(sanitized.contains("<p>Hi<a>x</a></p>"));
-    }
-
-    #[test]
-    fn keeps_safe_iframe_sources() {
-        let html =
-            r#"<iframe src="https://open.spotify.com/embed/track/a" loading="lazy"></iframe>"#;
-        let sanitized = sanitize_html(html, Some(&JsSanitizeOptions::default()));
-
-        assert!(sanitized.contains("<iframe"));
-        assert!(sanitized.contains("src=\"https://open.spotify.com/embed/track/a\""));
-    }
-
-    #[test]
-    fn keeps_safe_media_tags_and_attributes() {
-        let html = r#"<video controls muted loop playsinline poster="/poster.jpg" width="640" height="360" preload="metadata"><source src="/demo.webm" type="video/webm"><track src="/captions.vtt" kind="captions" srclang="en" label="English" default>Fallback</video><audio controls src="./clip.mp3"></audio><picture><source media="(min-width: 800px)" srcset="/hero-large.jpg 2x, /hero.jpg 1x" sizes="100vw"><img src="/hero.jpg" alt="Hero"></picture>"#;
-        let sanitized = sanitize_html(html, Some(&JsSanitizeOptions::default()));
-
-        assert!(sanitized.contains(
-            r#"<video controls muted loop playsinline poster="/poster.jpg" width="640" height="360" preload="metadata">"#
-        ));
-        assert!(sanitized.contains(r#"<source src="/demo.webm" type="video/webm">"#));
-        assert!(sanitized.contains(
-            r#"<track src="/captions.vtt" kind="captions" srclang="en" label="English" default>"#
-        ));
-        assert!(sanitized.contains("Fallback</video>"));
-        assert!(sanitized.contains(r#"<audio controls src="./clip.mp3"></audio>"#));
-        assert!(sanitized.contains(
-            r#"<source media="(min-width: 800px)" srcset="/hero-large.jpg 2x, /hero.jpg 1x" sizes="100vw">"#
-        ));
-    }
-
-    #[test]
-    fn removes_unsafe_media_urls() {
-        let html = r#"<video poster="javascript:alert(1)"><source src="javascript:alert(2)" srcset="javascript:alert(3) 1x, /safe.jpg 2x"></video><img src="/safe.jpg" srcset="/safe.jpg 1x, javascript:alert(4) 2x">"#;
-        let sanitized = sanitize_html(html, Some(&JsSanitizeOptions::default()));
-
-        assert!(!sanitized.contains("javascript"));
-        assert!(!sanitized.contains("poster="));
-        assert!(!sanitized.contains("srcset="));
-        assert!(sanitized.contains("<source>"));
-        assert!(sanitized.contains(r#"<img src="/safe.jpg">"#));
     }
 }
